- Fix Inconsistency Model STT - Rename Model class to SpeechModel - Rename ModelSet class to SpeechModelSet - Fix Aggregate Exception
