+++ src/IBM.WatsonDeveloperCloud.SpeechToText/v1/Model/SpeechModel.cs
@@ -1,0 +1,45 @@
+﻿/**
+* Copyright 2017 IBM Corp. All Rights Reserved.
+*
+* Licensed under the Apache License, Version 2.0 (the "License");
+* you may not use this file except in compliance with the License.
+* You may obtain a copy of the License at
+*
+*      http://www.apache.org/licenses/LICENSE-2.0
+*
+* Unless required by applicable law or agreed to in writing, software
+* distributed under the License is distributed on an "AS IS" BASIS,
+* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+* See the License for the specific language governing permissions and
+* limitations under the License.
+*
+*/
+
+using Newtonsoft.Json;
+
+namespace IBM.WatsonDeveloperCloud.SpeechToText.v1.Model
+{
+    public class SpeechModel
+    {
+        [JsonProperty("Name")]
+        public string Name { get; set; }
+
+        [JsonProperty("language")]
+        public string Language { get; set; }
+
+        [JsonProperty("rate")]
+        public int Rate { get; set; }
+
+        [JsonProperty("url")]
+        public string Url { get; set; }
+
+        [JsonProperty("supported_features")]
+        public SupportedFeatures SupportedFeatures { get; set; }
+
+        [JsonProperty("sessions")]
+        public string Sessions { get; set; }
+
+        [JsonProperty("description")]
+        public string Description { get; set; }
+    }
+}
