+++ src/IBM.WatsonDeveloperCloud.SpeechToText/v1/Model/SpeechRecognitionResult.cs
@@ -15,13 +15,24 @@
 *
 */
 
+using System.Collections.Generic;
 using Newtonsoft.Json;
 
-namespace IBM.WatsonDeveloperCloud.SpeechToText.v1.Models
+
+namespace IBM.WatsonDeveloperCloud.SpeechToText.v1.Model
 {
-    public class RecognizeStatus
+    public class SpeechRecognitionResult
     {
-        [JsonProperty("session")]
-        public SessionStatus Session { get; set; }
+        [JsonProperty("final")]
+        public bool Final { get; set; }
+
+        [JsonProperty("alternatives")]
+        public List<SpeechRecognitionAlternative> Alternatives { get; set; }
+
+        [JsonProperty("keywords_result")]
+        public KeywordResults KeywordResults { get; set; }
+
+        [JsonProperty("word_alternatives")]
+        public List<WordAlternativeResults> WordAlternativeResults { get; set; }
     }
 }
