--- src/IBM.WatsonDeveloperCloud.SpeechToText/v1/Models/KeywordResult.cs
@@ -1,36 +1,0 @@
-﻿/**
-* Copyright 2017 IBM Corp. All Rights Reserved.
-*
-* Licensed under the Apache License, Version 2.0 (the "License");
-* you may not use this file except in compliance with the License.
-* You may obtain a copy of the License at
-*
-*      http://www.apache.org/licenses/LICENSE-2.0
-*
-* Unless required by applicable law or agreed to in writing, software
-* distributed under the License is distributed on an "AS IS" BASIS,
-* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-* See the License for the specific language governing permissions and
-* limitations under the License.
-*
-*/
-
-using Newtonsoft.Json;
-
-namespace IBM.WatsonDeveloperCloud.SpeechToText.v1.Models
-{
-    public class KeywordResult
-    {
-        [JsonProperty("normalized_text")]
-        public string NormalizedText { get; set; }
-
-        [JsonProperty("start_time")]
-        public double StartTime { get; set; }
-
-        [JsonProperty("end_time")]
-        public double Endtime { get; set; }
-
-        [JsonProperty("confidence")]
-        public double Confidence { get; set; }
-    }
-}
