--- src/IBM.WatsonDeveloperCloud.SpeechToText/v1/Models/Session.cs
@@ -1,39 +1,0 @@
-﻿/**
-* Copyright 2017 IBM Corp. All Rights Reserved.
-*
-* Licensed under the Apache License, Version 2.0 (the "License");
-* you may not use this file except in compliance with the License.
-* You may obtain a copy of the License at
-*
-*      http://www.apache.org/licenses/LICENSE-2.0
-*
-* Unless required by applicable law or agreed to in writing, software
-* distributed under the License is distributed on an "AS IS" BASIS,
-* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-* See the License for the specific language governing permissions and
-* limitations under the License.
-*
-*/
-
-using Newtonsoft.Json;
-
-namespace IBM.WatsonDeveloperCloud.SpeechToText.v1.Models
-{
-    public class Session
-    {
-        [JsonProperty("session_id")]
-        public string SessionId { get; set; }
-
-        [JsonProperty("new_session_uri")]
-        public string NewSessionUri { get; set; }
-
-        [JsonProperty("recognize")]
-        public string Recognize { get; set; }
-
-        [JsonProperty("observe_result")]
-        public string ObserveResult { get; set; }
-
-        [JsonProperty("recognizeWS")]
-        public string RecognizeWS { get; set; }
-    }
-}
